Calibration data fixed. ADC data fixed. Sequential reads WIP.

+++ app/src/main/java/com/knobtviker/thermopile/data/sources/raw/bme680/Data.java
@@ -21,7 +21,7 @@ public class Data {
     public float temperature;
 
     // Pressure in Pascal
-    public int pressure = -1;
+    public float pressure = -1;
 
     // Humidity in % relative humidity x1000
     public float humidity;
